Create Component. - NavigationRail - TopAppBar - PermanentNavigationDrawer
