+++ Compose-My-Pohang-App/app/src/main/java/com/example/mypohangapp/component/MyPohangTopAppBar.kt
@@ -1,0 +1,45 @@
+package com.example.mypohangapp.component
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowBack
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import com.example.mypohangapp.R
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun MyPohangTopAppBar(
+    modifier: Modifier = Modifier,
+    currentTitle: Int,
+    canNavigateBack: Boolean,
+    navigateUp: () -> Unit
+) {
+    TopAppBar(
+        title = { Text(stringResource(id = currentTitle)) },
+        modifier = modifier,
+        colors = TopAppBarDefaults.topAppBarColors(
+            containerColor = MaterialTheme.colorScheme.primaryContainer
+        ),
+        navigationIcon = {
+            if (canNavigateBack) {
+                IconButton(onClick = navigateUp) {
+                    Icon(
+                        imageVector = Icons.Filled.ArrowBack,
+                        contentDescription = stringResource(R.string.back_button)
+                    )
+                }
+            } else {
+                Box {}
+            }
+        }
+    )
+}
